// src/generated/schemas/authSignInDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * GREGORI API
 * GREGORI 종합 쇼핑몰 REST API 명세서
 * OpenAPI spec version: 1.0.0
 */

export interface AuthSignInDto {
  email: string;
  /** @pattern ^(?=(.*[a-zA-Z].*){2,})(?=.*\d.*)(?=.*\W.*)[a-zA-Z0-9\S]{8,15}$ */
  password: string;
}